Fix incorrect types

+++ network/CNetGamePlayerDataMsg.hpp
@@ -24,16 +24,16 @@ namespace rage {
 class CNetGamePlayerDataMsg : public rage::playerDataMsg
 {
 public:
-    uint32_t m_account_id; //0x0010
+    int32_t m_account_id; //0x0010
     uint32_t m_player_type; //0x0014
     uint32_t m_matchmaking_group; //0x0018
     uint32_t m_flags; //0x001C
-    uint32_t m_team; //0x0020
-    uint64_t m_crew_id; //0x0028
-    uint32_t m_aim_preference; //0x0030
+    int32_t m_team; //0x0020
+    int64_t m_crew_id; //0x0028
+    int32_t m_aim_preference; //0x0030
     uint16_t m_rank; //0x0034
     uint16_t m_rating; //0x0036
-    uint16_t m_region; //0x0038
+    uint32_t m_region; //0x0038
 }; //Size: 0x0040
 static_assert(sizeof(CNetGamePlayerDataMsg) == 0x40);
 #pragma pack(pop)
